Rewrite of TVP support; Big conversion cleanup
Uses IEnumerable<SqlDataRecord> instead of DbDataReader.
This allows SqlServer to look at the metadata on the type.
Unfortunately, this is a lower-level API so we need to do better conversions.
The updated conversion code should support even more cases.

In the end, this adds support for TVPs with identity columns.

Implements #255 and fixes #308.